+++ main.py
@@ -33,3 +33,18 @@ print()
 print("Welcome to Project NOVA,", name)
 
 print("=" * 50)
+
+
+print("\n=== STUDY TRACKER ===")
+
+python_hours = float(input("Python hours today: "))
+math_hours = float(input("Math hours today: "))
+ai_hours = float(input("AI hours today: "))
+
+total_hours = python_hours + math_hours + ai_hours
+
+print("\nToday's Summary")
+print("Python:", python_hours)
+print("Math:", math_hours)
+print("AI:", ai_hours)
+print("Total Hours:", total_hours)