+++ mini-projects/day-04-operators.py
@@ -1,0 +1,11 @@
+a = 15
+b = 20
+print(a+b)
+print(b-a)
+print(a*b)
+print(a%b)
+print(a//b)
+print(a/b)
+
+age = int(input("Enter your age :"))
+print(age)