+++ mini-projects/projects/day04-project-operator.py
@@ -1,0 +1,4 @@
+weight = float(input("Enter the weight :"))
+height = float(input("Enter the height :"))
+BMI = weight/(height**2)
+print("Your BMI is :",BMI)